split off more code from views file
