+++ flaim/reports/views.py
@@ -1,22 +1,18 @@
 import re
 from functools import partial
 from operator import is_not
-from textwrap import wrap
 from urllib.parse import unquote
 
 import numpy as np
 import pandas as pd
-import plotly.figure_factory as ff
-import plotly.graph_objects as go
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.views.generic import TemplateView
-from numpy.linalg import LinAlgError
-from plotly.express.colors import qualitative
-from plotly.io import to_html
 
 from flaim.database.product_mappings import PRODUCT_STORES, REFERENCE_CATEGORIES_DICT, \
     REFERENCE_SUBCATEGORIES_CODING_DICT
 from flaim.reports.data import ReportData, StoreReportData
+from flaim.reports.util import get_nutrient_color, get_rank_suffix
+from flaim.reports.plots import nutrient_distribution_plot, category_nutrient_distribution_plot
 
 
 class ProductView(LoginRequiredMixin, TemplateView):
@@ -98,7 +94,7 @@ class SubcategoryView(LoginRequiredMixin, TemplateView):
         context['sugar_median'] = f'{plot_df.sugar.median() * 100:.0f}%'
 
         # Visualizations
-        context['figure1'] = get_nutrient_distribution_plot(plot_df)
+        context['figure1'] = nutrient_distribution_plot(plot_df)
         return context
 
 
@@ -109,19 +105,18 @@ class CategoryView(LoginRequiredMixin, TemplateView):
         context = super().get_context_data(**kwargs)
         data = ReportData()
         df = data.df.copy()  # temp copy
-        plot_df = df.drop_duplicates(subset='name')
 
         context['product_categories'] = REFERENCE_CATEGORIES_DICT.keys()
 
         # set default category
         if 'category' not in self.kwargs:
-            context['category'] = plot_df['category_text'].value_counts().index[0]
+            context['category'] = df['category_text'].value_counts().index[0]
         else:
             # pulls category from URL e.g. /reports/category/Beverages
             context['category'] = unquote(self.kwargs['category'])
 
-        context['category_count'] = plot_df['category_text'].nunique()
-        medians = plot_df.groupby('category_text').median()
+        context['category_count'] = df['category_text'].nunique()
+        medians = df.groupby('category_text').median()
 
         context['sodium_rank'] = get_rank_suffix(medians.sort_values(by='sodium_dv', ascending=False)
                                                  .index.get_loc(context['category']))
@@ -130,14 +125,14 @@ class CategoryView(LoginRequiredMixin, TemplateView):
         context['sugar_rank'] = get_rank_suffix(medians.sort_values(by='sugar', ascending=False)
                                                 .index.get_loc(context['category']))
 
-        plot_df = plot_df.loc[plot_df['category_text'] == context['category']]
+        df = df.loc[df['category_text'] == context['category']]
 
-        ingredient_count = plot_df['ingredients'].str.findall(',').fillna('').apply(lambda row: len(row) + 1)
+        ingredient_count = df['ingredients'].str.findall(',').fillna('').apply(lambda row: len(row) + 1)
         context['ingredient_q25'] = int(ingredient_count.quantile(0.25))
         context['ingredient_q75'] = int(ingredient_count.quantile(0.75))
 
         # This gets the 3 most common ingredients but does not preprocess the ingredient lists
-        common_ingredients = pd.Series(' '.join(plot_df['ingredients'].fillna('').str.lower().tolist())
+        common_ingredients = pd.Series(' '.join(df['ingredients'].fillna('').str.lower().tolist())
                                        .split(',')).value_counts().head(3).index.tolist()
         if len(common_ingredients) >= 3:
             context['common_ingredients'] = f'The three most common ingredients in this category are \
@@ -152,24 +147,24 @@ class CategoryView(LoginRequiredMixin, TemplateView):
 
         # Top bar
         context['image'] = context['category'].lower()
-        context['product_count'] = plot_df.shape[0]
-        context['store_count'] = plot_df['store'].nunique()
-        context['manual_category_count'] = plot_df['manual_category_text'].dropna().shape[0]
+        context['product_count'] = df.shape[0]
+        context['store_count'] = df['store'].nunique()
+        context['manual_category_count'] = df['manual_category_text'].dropna().shape[0]
         context['predicted_category_count'] = context['product_count'] - context['manual_category_count']
 
-        context['calorie_median'] = f'{plot_df.calories.median():.0f}'
+        context['calorie_median'] = f'{df.calories.median():.0f}'
 
-        context['sodium_color'] = get_nutrient_color(plot_df.sodium_dv.median())
-        context['sodium_median'] = f'{plot_df.sodium_dv.median() * 100:.0f}%'
+        context['sodium_color'] = get_nutrient_color(df.sodium_dv.median())
+        context['sodium_median'] = f'{df.sodium_dv.median() * 100:.0f}%'
 
-        context['fat_color'] = get_nutrient_color(plot_df.saturatedfat_dv.median())
-        context['fat_median'] = f'{plot_df.saturatedfat_dv.median() * 100:.0f}%'
+        context['fat_color'] = get_nutrient_color(df.saturatedfat_dv.median())
+        context['fat_median'] = f'{df.saturatedfat_dv.median() * 100:.0f}%'
 
-        context['sugar_color'] = get_nutrient_color(plot_df.sugar.median())
-        context['sugar_median'] = f'{plot_df.sugar.median() * 100:.0f}%'
+        context['sugar_color'] = get_nutrient_color(df.sugar.median())
+        context['sugar_median'] = f'{df.sugar.median() * 100:.0f}%'
 
         # Visualizations
-        context['figure1'] = get_nutrient_distribution_plot(plot_df)
+        context['figure1'] = nutrient_distribution_plot(df)
         return context
 
 
@@ -279,134 +274,5 @@ class StoreView(LoginRequiredMixin, TemplateView):
         context['complete'] = complete_df['complete'].value_counts()[True]
 
         # Visualizations
-        context['figure1'] = get_category_nutrient_distribution_plot(df)
+        context['figure1'] = category_nutrient_distribution_plot(df)
         return context
-
-
-def get_nutrient_color(value):
-    return 'text-success' if value <= 0.15 else 'text-danger'
-
-
-def get_rank_suffix(i):
-    i = i + 1
-    if i >= 11 and 11 <= int(str(i)[-2:]) <= 13:
-        return f'{i}th'
-    remainder = i % 10
-    if remainder == 1:
-        return f'{i}st'
-    elif remainder == 2:
-        return f'{i}nd'
-    elif remainder == 3:
-        return f'{i}rd'
-    else:
-        return f'{i}th'
-
-
-# Figure generation
-def get_nutrient_distribution_plot(df):
-    nutrients = ['sodium_dv', 'saturatedfat_dv', 'sugar']
-
-    try:
-        fig = ff.create_distplot(df[nutrients].dropna(how='all').fillna(0).T.values,
-                                 ['Sodium', 'Saturated Fat', 'Sugar'], colors=qualitative.Vivid, histnorm='probability',
-                                 bin_size=0.01, show_rug=False, )
-        fig.update_layout(
-            width=1100,
-            font_size=18,
-            xaxis=dict(
-                title='Daily Value',
-                tickformat='%',
-                showgrid=True,
-                range=[0, min([1, max(df[n].quantile(0.95) for n in nutrients)])]
-            ),
-            yaxis=dict(
-                tickformat='%',
-                title='Proportion of Products'
-            ),
-            margin=dict(
-                l=100,
-                r=20,
-                b=30,
-                t=30,
-            ),
-            legend_title='Nutrients'
-        )
-
-        fig.add_shape(dict(
-            type='line',
-            yref='paper',
-            x0=0.15,
-            y0=0,
-            x1=0.15,
-            y1=1,
-            line=dict(
-                color='Black',
-                width=2
-            )))
-
-        fig.add_annotation(text='← Low in Nutrient',
-                           yref='paper',
-                           x=0.145, y=1,
-                           showarrow=False,
-                           xanchor='right',
-                           yanchor='bottom',
-                           font_color='green')
-        fig.add_annotation(text='High in Nutrient →',
-                           yref='paper',
-                           x=0.155, y=1,
-                           showarrow=False,
-                           xanchor='left',
-                           yanchor='bottom',
-                           font_color='red')
-    except LinAlgError:
-        return "Graph can't be displayed."
-    except ValueError:
-        return "Not enough data to generate graph."
-
-    return to_html(fig, include_plotlyjs=False, full_html=False)
-
-
-def get_category_nutrient_distribution_plot(df):
-    def over_15(row):
-        return 1 if row > 0.15 else 0
-
-    nutrients = ['sodium_dv', 'saturatedfat_dv', 'sugar']
-    nutrient_map = {'sodium_dv': 'Sodium', 'saturatedfat_dv': 'Saturated Fat', 'sugar': 'Sugar'}
-    categories = ['Beverages', 'Cereals and Other Grain Products', 'Dairy Products and Substitutes',
-                  'Meat and Poultry, Products and Substitutes', 'Snacks']
-
-    category_labels = ['<br>'.join(wrap(c, 20)) for c in categories]
-
-    plot_df = df.loc[df['category_text'].isin(categories)]
-
-    for n in nutrients:
-        plot_df[f'{n}_count'] = plot_df[n].apply(lambda row: over_15(row))
-
-    plot_df = plot_df[['category_text'] + [f'{n}_count' for n in nutrients]].groupby(
-        'category_text').sum().reindex(categories)
-
-    data = [
-        go.Bar(name=nutrient_map[nutrient], x=category_labels, y=plot_df[f'{nutrient}_count']) for nutrient in nutrients
-    ]
-
-    layout = dict(
-        width=1100,
-        font_size=18,
-        margin=dict(
-            l=100,
-            r=20,
-            b=20,
-            t=30,
-        ),
-        yaxis_title='Products Exceeding Threshold',
-        xaxis=dict(
-            title='Food Category',
-            showgrid=True,
-            tickson='boundaries',
-            tickangle=0,
-        ),
-    )
-
-    fig = go.Figure(data=data, layout=layout)
-
-    return to_html(fig, include_plotlyjs=False, full_html=False)
